requirements.txt
alembic==1.13.0 ; python_version >= "3.11" and python_version < "3.12"
blinker==1.7.0 ; python_version >= "3.11" and python_version < "3.12"
bootstrap-flask==2.3.3 ; python_version >= "3.11" and python_version < "3.12"
certifi==2023.11.17 ; python_version >= "3.11" and python_version < "3.12"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.11" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.12" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==41.0.7 ; python_version >= "3.11" and python_version < "3.12"
deprecated==1.2.14 ; python_version >= "3.11" and python_version < "3.12"
flask-wtf==1.2.1 ; python_version >= "3.11" and python_version < "3.12"
flask==2.3.3 ; python_version >= "3.11" and python_version < "3.12"
gitdb==4.0.11 ; python_version >= "3.11" and python_version < "3.12"
gitpython==3.1.40 ; python_version >= "3.11" and python_version < "3.12"
greenlet==3.0.1 ; python_version >= "3.11" and python_version < "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
gunicorn==21.2.0 ; python_version >= "3.11" and python_version < "3.12"
idna==3.6 ; python_version >= "3.11" and python_version < "3.12"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "3.12"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "3.12"
lizard==1.17.10 ; python_version >= "3.11" and python_version < "3.12"
loguru==0.7.2 ; python_version >= "3.11" and python_version < "3.12"
mako==1.3.0 ; python_version >= "3.11" and python_version < "3.12"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "3.12"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "3.12"
mypy==1.7.1 ; python_version >= "3.11" and python_version < "3.12"
packaging==23.2 ; python_version >= "3.11" and python_version < "3.12"
psycopg-binary==3.1.14 ; implementation_name != "pypy" and python_version >= "3.11" and python_version < "3.12"
psycopg-pool==3.2.0 ; python_version >= "3.11" and python_version < "3.12"
psycopg[binary,pool]==3.1.14 ; python_version >= "3.11" and python_version < "3.12"
pycparser==2.21 ; python_version >= "3.11" and python_version < "3.12"
pydriller==2.6 ; python_version >= "3.11" and python_version < "3.12"
pygithub==2.1.1 ; python_version >= "3.11" and python_version < "3.12"
pyjwt[crypto]==2.8.0 ; python_version >= "3.11" and python_version < "3.12"
pynacl==1.5.0 ; python_version >= "3.11" and python_version < "3.12"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "3.12"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "3.12"
python-gitlab==4.2.0 ; python_version >= "3.11" and python_version < "3.12"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "3.12"
requests-toolbelt==1.0.0 ; python_version >= "3.11" and python_version < "3.12"
requests==2.31.0 ; python_version >= "3.11" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.11" and python_version < "3.12"
smmap==5.0.1 ; python_version >= "3.11" and python_version < "3.12"
sqlalchemy-utils==0.41.1 ; python_version >= "3.11" and python_version < "3.12"
sqlalchemy==2.0.23 ; python_version >= "3.11" and python_version < "3.12"
sqlalchemy[mypy]==2.0.23 ; python_version >= "3.11" and python_version < "3.12"
tqdm==4.66.1 ; python_version >= "3.11" and python_version < "3.12"
types-pytz==2023.3.1.1 ; python_version >= "3.11" and python_version < "3.12"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "3.12"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "3.12" and sys_platform == "win32"
urllib3==2.1.0 ; python_version >= "3.11" and python_version < "3.12"
werkzeug==3.0.1 ; python_version >= "3.11" and python_version < "3.12"
win32-setctime==1.1.0 ; python_version >= "3.11" and python_version < "3.12" and sys_platform == "win32"
wrapt==1.16.0 ; python_version >= "3.11" and python_version < "3.12"
wtforms==3.1.1 ; python_version >= "3.11" and python_version < "3.12"
